@@ -14,31 +14,35 @@ description: Flutter version of Space Mapper with 2020 upgrades
 version: 3.0.2+18
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_background_geolocation: '^1.3.2'
-  shared_preferences: ^0.5.6
-  background_fetch: ^0.5.3
-  url_launcher: ^5.2.5
-  http: ^0.12.1
-  flutter_map: ^0.9.0
-  flutter_webview_plugin: ^0.3.11
-  sqflite: ^1.3.0
-  path_provider: ^1.6.7
-  share: ^0.6.4+1
-  uuid: 2.0.4
-  flutter_form_builder: ^3.9.0
+  flutter_background_geolocation: ^4.3.0
+  shared_preferences: ^2.0.8
+  background_fetch: ^1.0.1
+  url_launcher: ^6.0.12
+  http: ^0.13.4
+  flutter_map: ^0.14.0
+  # flutter_webview_plugin: ^0.4.0
+  webview_flutter: ^2.1.1
+  sqflite: ^2.0.0+4
+  path_provider: ^2.0.5
+  share: ^2.0.4
+  uuid: 3.0.5
+  flutter_form_builder: ^6.1.0+1
+  latlong2: ^0.8.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  test: ^1.17.10
+  network_image_mock: ^2.0.1
 
 
 # For information on the generic Dart part of this file, see the
