@@ -2,6 +2,7 @@ import 'package:asm/ui/form_view.dart';
 import 'package:asm/ui/side_drawer.dart';
 import 'package:asm/util/env.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'dart:async';
 import 'dart:convert';
 
@@ -22,31 +23,36 @@ JsonEncoder encoder = new JsonEncoder.withIndent("     ");
 /// The main home-screen of the AdvancedApp.  Builds the Scaffold of the App.
 ///
 class HomeView extends StatefulWidget {
+  HomeView(this.appName);
+  final String appName;
+
   @override
-  State createState() => HomeViewState();
+  State createState() => HomeViewState(appName);
 }
 
 class HomeViewState extends State<HomeView>
     with TickerProviderStateMixin<HomeView>, WidgetsBindingObserver {
   Future<SharedPreferences> _prefs = SharedPreferences.getInstance();
-  TabController _tabController;
+  //late TabController _tabController;
 
-  bool _isMoving;
-  bool _enabled;
-  String _motionActivity;
-  String _odometer;
+  late String appName;
+  //late bool _isMoving;
+  late bool _enabled;
+  //late String _motionActivity;
+  //late String _odometer;
+
+  HomeViewState(this.appName);
 
   @override
   void initState() {
     super.initState();
 
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
 
-    _isMoving = false;
+    //_isMoving = false;
     _enabled = true;
-    _motionActivity = 'UNKNOWN';
-    _odometer = '0';
-
+    //_motionActivity = 'UNKNOWN';
+    //_odometer = '0';
 
     initPlatformState();
   }
@@ -54,26 +60,28 @@ class HomeViewState extends State<HomeView>
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
     print("[home_view didChangeAppLifecycleState] : $state");
+    //TODO: The interior of these 'if statements' is empty
     if (state == AppLifecycleState.paused) {
     } else if (state == AppLifecycleState.resumed) {}
   }
 
   void initPlatformState() async {
-
     SharedPreferences prefs = await _prefs;
-    String sample_id = prefs.getString("sample_id");
-    String user_uuid = prefs.getString("user_uuid");
+    // ignore: non_constant_identifier_names
+    String? sample_id = prefs.getString("sample_id");
+    // ignore: non_constant_identifier_names
+    String? user_uuid = prefs.getString("user_uuid");
 
     if (sample_id == null || user_uuid == null) {
       prefs.setString("user_uuid", Uuid().v4());
       prefs.setString("sample_id", ENV.DEFAULT_SAMPLE_ID);
     }
 
-
     _configureBackgroundGeolocation(user_uuid, sample_id);
     _configureBackgroundFetch();
   }
 
+  // ignore: non_constant_identifier_names
   void _configureBackgroundGeolocation(user_uuid, sample_id) async {
     // 1.  Listen to events (See docs for all 13 available events).
     bg.BackgroundGeolocation.onLocation(_onLocation, _onLocationError);
@@ -113,17 +121,16 @@ class HomeViewState extends State<HomeView>
         .then((bg.State state) {
       print('[ready] ${state.toMap()}');
 
-      if (state.schedule.isNotEmpty) {
+      if (state.schedule!.isNotEmpty) {
         bg.BackgroundGeolocation.startSchedule();
       }
       setState(() {
         _enabled = state.enabled;
-        _isMoving = state.isMoving;
+        //_isMoving = state.isMoving!;
       });
     }).catchError((error) {
       print('[ready] ERROR: $error');
     });
-
   }
 
   // Configure BackgroundFetch (not required by BackgroundGeolocation).
@@ -144,13 +151,12 @@ class HomeViewState extends State<HomeView>
       SharedPreferences prefs = await SharedPreferences.getInstance();
       int count = 0;
       if (prefs.get("fetch-count") != null) {
-        count = prefs.getInt("fetch-count");
+        count = prefs.getInt("fetch-count")!;
       }
       prefs.setInt("fetch-count", ++count);
       print('[BackgroundFetch] count: $count');
 
       // Test scheduling a custom-task in fetch event.
-
 
       if (taskId == 'flutter_background_fetch') {
         BackgroundFetch.scheduleTask(TaskConfig(
@@ -180,7 +186,7 @@ class HomeViewState extends State<HomeView>
         print('[start] success: $state');
         setState(() {
           _enabled = state.enabled;
-          _isMoving = state.isMoving;
+          //_isMoving = state.isMoving!;
         });
       };
       bg.State state = await bg.BackgroundGeolocation.state;
@@ -194,7 +200,7 @@ class HomeViewState extends State<HomeView>
         print('[stop] success: $state');
         setState(() {
           _enabled = state.enabled;
-          _isMoving = state.isMoving;
+          //_isMoving = state.isMoving!;
         });
       };
       bg.BackgroundGeolocation.stop().then(callback);
@@ -202,7 +208,7 @@ class HomeViewState extends State<HomeView>
   }
 
   // Manually toggle the tracking state:  moving vs stationary
-  void _onClickChangePace() {
+  /*void _onClickChangePace() {
     setState(() {
       _isMoving = !_isMoving;
     });
@@ -215,7 +221,7 @@ class HomeViewState extends State<HomeView>
     });
 
     if (!_isMoving) {}
-  }
+  }*/
 
   // Manually fetch the current position.
   void _onClickGetCurrentPosition() async {
@@ -245,7 +251,7 @@ class HomeViewState extends State<HomeView>
     print('[${bg.Event.LOCATION}] - $location');
 
     setState(() {
-      _odometer = (location.odometer / 1000.0).toStringAsFixed(1);
+      //_odometer = (location.odometer / 1000.0).toStringAsFixed(1);
     });
   }
 
@@ -257,14 +263,14 @@ class HomeViewState extends State<HomeView>
   void _onMotionChange(bg.Location location) {
     print('[${bg.Event.MOTIONCHANGE}] - $location');
     setState(() {
-      _isMoving = location.isMoving;
+      //_isMoving = location.isMoving;
     });
   }
 
   void _onActivityChange(bg.ActivityChangeEvent event) {
     print('[${bg.Event.ACTIVITYCHANGE}] - $event');
     setState(() {
-      _motionActivity = event.activity;
+      //_motionActivity = event.activity;
     });
   }
 
@@ -296,20 +302,19 @@ class HomeViewState extends State<HomeView>
       // Execute an HTTP request to test an async operation completes.
       String url = "${ENV.TRACKER_HOST}/api/devices";
       bg.State state = await bg.BackgroundGeolocation.state;
-      http.read(url, headers: {
-        "Authorization": "Bearer ${state.authorization.accessToken}"
+      http.read(Uri.parse(url), headers: {
+        "Authorization": "Bearer ${state.authorization!.accessToken}"
       }).then((String result) {
         print("[http test] success: $result");
-        bg.BackgroundGeolocation.playSound(util.Dialog.getSoundId("TEST_MODE_CLICK"));
+        bg.BackgroundGeolocation.playSound(
+            util.Dialog.getSoundId("TEST_MODE_CLICK"));
         bg.BackgroundGeolocation.stopBackgroundTask(taskId);
       }).catchError((dynamic error) {
         print("[http test] failed: $error");
         bg.BackgroundGeolocation.stopBackgroundTask(taskId);
       });
     });
-
   }
-
 
   void _onSchedule(bg.State state) {
     print('[${bg.Event.SCHEDULE}] - $state');
@@ -343,10 +348,11 @@ class HomeViewState extends State<HomeView>
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Space Mapper'),
+        title: Text(appName),
         centerTitle: true,
         backgroundColor: Colors.blueGrey,
-        brightness: Brightness.light,
+        systemOverlayStyle:
+            SystemUiOverlayStyle(statusBarBrightness: Brightness.light),
         leading: Builder(
           builder: (BuildContext context) {
             return IconButton(
@@ -358,14 +364,19 @@ class HomeViewState extends State<HomeView>
             );
           },
         ),
-        actions: <Widget>[/*
+        actions: <Widget>[
+          /*
           IconButton(
             icon: Icon(Icons.gps_fixed),
             color: Colors.yellow,
             onPressed:
               _onClickGetCurrentPosition,
           ),*/
-          Switch(value: _enabled, onChanged: _onClickEnable, activeColor: Colors.yellow,),
+          Switch(
+            value: _enabled,
+            onChanged: _onClickEnable,
+            activeColor: Colors.yellow,
+          ),
         ],
       ),
       //body: body,
@@ -373,8 +384,9 @@ class HomeViewState extends State<HomeView>
       body: MapView(),
       floatingActionButton: FloatingActionButton(
         onPressed: () {
-     //     Navigator.push(context, MaterialPageRoute(builder: (context) => FormView()));
-          _onClickGetCurrentPosition;
+          Navigator.push(
+              context, MaterialPageRoute(builder: (context) => FormView()));
+          _onClickGetCurrentPosition();
         },
         child: Icon(Icons.gps_fixed),
         backgroundColor: Colors.blue,
@@ -386,6 +398,4 @@ class HomeViewState extends State<HomeView>
   void dispose() {
     super.dispose();
   }
-
 }
-
