@@ -1,19 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_background_geolocation/flutter_background_geolocation.dart' as bg;
+import 'package:flutter_background_geolocation/flutter_background_geolocation.dart'
+    as bg;
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/plugin_api.dart';
-import 'package:latlong/latlong.dart';
+import 'package:latlong2/latlong.dart';
 
 class MapView extends StatefulWidget {
   @override
   State createState() => MapViewState();
 }
 
-class MapViewState extends State<MapView> with AutomaticKeepAliveClientMixin<MapView> {
+class MapViewState extends State<MapView>
+    with AutomaticKeepAliveClientMixin<MapView> {
   @override
-  bool get wantKeepAlive { return true; }
+  bool get wantKeepAlive {
+    return true;
+  }
 
-  bg.Location _stationaryLocation;
+  late bg.Location _stationaryLocation;
 
   List<CircleMarker> _currentPosition = [];
   List<LatLng> _polyline = [];
@@ -22,18 +26,17 @@ class MapViewState extends State<MapView> with AutomaticKeepAliveClientMixin<Map
   List<Polyline> _motionChangePolylines = [];
   List<CircleMarker> _stationaryMarker = [];
 
-
   LatLng _center = new LatLng(51.5, -0.09);
-  MapController _mapController;
-  MapOptions _mapOptions;
+  late MapController _mapController;
+  late MapOptions _mapOptions;
 
   @override
   void initState() {
     super.initState();
     _mapOptions = new MapOptions(
-        onPositionChanged: _onPositionChanged,
-        center: _center,
-        zoom: 16.0,
+      onPositionChanged: _onPositionChanged,
+      center: _center,
+      zoom: 16.0,
     );
     _mapController = new MapController();
 
@@ -41,16 +44,14 @@ class MapViewState extends State<MapView> with AutomaticKeepAliveClientMixin<Map
     bg.BackgroundGeolocation.onMotionChange(_onMotionChange);
     bg.BackgroundGeolocation.onEnabledChange(_onEnabledChange);
 
-
-    _mapController.onReady.then((_){
+    _mapController.onReady.then((_) {
       _displayStoredLocations();
     });
-
   }
 
   void _displayStoredLocations() async {
     List allLocations = await bg.BackgroundGeolocation.locations;
-    for (var thisLocation in allLocations){
+    for (var thisLocation in allLocations) {
       _onLocation(bg.Location(thisLocation));
     }
   }
@@ -61,7 +62,7 @@ class MapViewState extends State<MapView> with AutomaticKeepAliveClientMixin<Map
 //      _polyline.clear();
 //      _stopLocations.clear();
 //      _motionChangePolylines.clear();
- //     _stationaryMarker.clear();
+      //     _stationaryMarker.clear();
     }
   }
 
@@ -76,21 +77,22 @@ class MapViewState extends State<MapView> with AutomaticKeepAliveClientMixin<Map
     _stationaryMarker.clear();
 
     if (location.isMoving) {
-      if (_stationaryLocation == null) {
-        _stationaryLocation = location;
-      }
+      //if (_stationaryLocation == null) { //TODO: The operand can't be null, so the condition is always false
+      _stationaryLocation = location;
+      //}
       // Add previous stationaryLocation as a small red stop-circle.
       _stopLocations.add(_buildStopCircleMarker(_stationaryLocation));
       // Create the green motionchange polyline to show where tracking engaged from.
-      _motionChangePolylines.add(_buildMotionChangePolyline(_stationaryLocation, location));
+      _motionChangePolylines
+          .add(_buildMotionChangePolyline(_stationaryLocation, location));
     } else {
       // Save a reference to the location where we became stationary.
       _stationaryLocation = location;
       // Add the big red stationaryRadius circle.
 
       // TAKING THIS OUT FOR NOW
- //      bg.State state = await bg.BackgroundGeolocation.state;
- //     _stationaryMarker.add(_buildStationaryCircleMarker(location, state));
+      //      bg.State state = await bg.BackgroundGeolocation.state;
+      //     _stationaryMarker.add(_buildStationaryCircleMarker(location, state));
     }
   }
 
@@ -100,23 +102,19 @@ class MapViewState extends State<MapView> with AutomaticKeepAliveClientMixin<Map
 
     _updateCurrentPositionMarker(ll);
 
-    if (location.sample) { return; }
+    if (location.sample != null) {
+      if (location.sample == true) {
+        return;
+      }
+    }
 
     // Add a point to the tracking polyline.
     _polyline.add(ll);
     // Add a marker for the recorded location.
     //_locations.add(_buildLocationMarker(location));
-    _locations.add(CircleMarker(
-        point: ll,
-        color: Colors.black,
-        radius: 5.0
-    ));
+    _locations.add(CircleMarker(point: ll, color: Colors.black, radius: 5.0));
 
-    _locations.add(CircleMarker(
-        point: ll,
-        color: Colors.blue,
-        radius: 4.0
-    ));
+    _locations.add(CircleMarker(point: ll, color: Colors.blue, radius: 4.0));
   }
 
   /// Update Big Blue current position dot.
@@ -124,49 +122,38 @@ class MapViewState extends State<MapView> with AutomaticKeepAliveClientMixin<Map
     _currentPosition.clear();
 
     // White background
-    _currentPosition.add(CircleMarker(
-        point: ll,
-        color: Colors.white,
-        radius: 10
-    ));
+    _currentPosition
+        .add(CircleMarker(point: ll, color: Colors.white, radius: 10));
     // Blue foreground
-    _currentPosition.add(CircleMarker(
-        point: ll,
-        color: Colors.blue,
-        radius: 7
-    ));
+    _currentPosition
+        .add(CircleMarker(point: ll, color: Colors.blue, radius: 7));
   }
 
-  CircleMarker _buildStationaryCircleMarker(bg.Location location, bg.State state) {
+  /*CircleMarker _buildStationaryCircleMarker(
+      bg.Location location, bg.State state) {
     return new CircleMarker(
         point: LatLng(location.coords.latitude, location.coords.longitude),
         color: Color.fromRGBO(255, 0, 0, 0.5),
         useRadiusInMeter: true,
-        radius: (state.trackingMode == 1) ? 200 : (state.geofenceProximityRadius / 2)
-    );
-  }
+        radius: (state.trackingMode == 1)
+            ? 200
+            : (state.geofenceProximityRadius! / 2));
+  }*/
 
   Polyline _buildMotionChangePolyline(bg.Location from, bg.Location to) {
-    return new Polyline(
-        points: [
-          LatLng(from.coords.latitude, from.coords.longitude),
-          LatLng(to.coords.latitude, to.coords.longitude)
-        ],
-        strokeWidth: 10.0,
-        color: Color.fromRGBO(22, 190, 66, 0.7)
-    );
+    return new Polyline(points: [
+      LatLng(from.coords.latitude, from.coords.longitude),
+      LatLng(to.coords.latitude, to.coords.longitude)
+    ], strokeWidth: 10.0, color: Color.fromRGBO(22, 190, 66, 0.7));
   }
-
 
   CircleMarker _buildStopCircleMarker(bg.Location location) {
     return new CircleMarker(
         point: LatLng(location.coords.latitude, location.coords.longitude),
         color: Color.fromRGBO(200, 0, 0, 0.3),
         useRadiusInMeter: false,
-        radius: 20
-    );
+        radius: 20);
   }
-
 
   void _onPositionChanged(MapPosition pos, bool hasGesture) {
     _mapOptions.crs.scale(_mapController.zoom);
@@ -179,11 +166,9 @@ class MapViewState extends State<MapView> with AutomaticKeepAliveClientMixin<Map
       mapController: _mapController,
       options: _mapOptions,
       layers: [
-
         new TileLayerOptions(
-              urlTemplate: "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
-              subdomains: ['a', 'b', 'c']
-        ),
+            urlTemplate: "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
+            subdomains: ['a', 'b', 'c']),
         new PolylineLayerOptions(
           polylines: [
             new Polyline(
@@ -206,4 +191,3 @@ class MapViewState extends State<MapView> with AutomaticKeepAliveClientMixin<Map
     );
   }
 }
-
